Wait for background command to finish before destroying
The `senjo::Thread` class's destructor has a call to std::thread::join to
make sure its underlying thread is finished before destruction.

The `senjo::Thread::staticRun` method calls a virtual method of a child
class of the Thread to perform some work. For example, `goCommand` is a
child class of Thread, and the `doWork` method performs the engine search.

`senjo::Thread::execute` creates a `std::thread` and has it execute
`senjo::Thread::staticRun`

When reading commands from a file, the main method can exit very
quickly, and this can result in the `UCIAdapter` destructor being
called immediately after `senjo::Thread::execute`, which calls the
`goCommand` destructor. On my M1 mac, this sometimes happens before
the `std::thread` has started executing the `staticRun` method.
Because of the call to `std::thread::join` in the destructor, the
destructor will pause while it waits for the thread to run its `staticRun`
command.

As near as I can tell, the child's overload of `doWork` is not available
because the object is in the process of being destroyed, and the app
crashes because it is now attempting to call a pure virtual method.

This commit waits for the background thread to finish in the `UCIAdapter`
destructor. The `std::thread` can safely call `senjo::Thread::doWork`
since its destructor has not been invoked yet.